ADDED: validateBody
