+++ controllers/contacts.js
@@ -1,15 +1,9 @@
-const Joi = require("joi");
+
 
 const contacts = require("../models/contacts");
 const { HttpError, ctrlWrapper } = require("../helpers");
-const { addContact } = require("../models/contacts");
 
 
-const addShema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
 
 const getAll = async ( req, res, next) => { 
   
@@ -31,11 +25,7 @@ const getById = async (req, res, next) => {
 
 const add = async (req, res, next) => { 
 
-    const { error } = addShema.validate(req.body);
-    if (error) {
-      throw HttpError(error, error.message);
-    }
-    const result = await addContact(req.body);
+    const result = await contacts.addContact(req.body);
     res.status(201).json(result);
   
 }
@@ -52,10 +42,6 @@ const deleteById = async (req, res, next) => {
 }
 const updateById = async (req, res, next) => { 
 
-    const { error } = addShema.validate(req.body);
-    if (error) {
-      throw HttpError(error, error.message);
-    }
     const { contactId } = req.params;
     const result = await contacts.updateContact(contactId, req.body);
     if (!result) {
