+++ middelwares/index.js
@@ -1,0 +1,3 @@
+const validateBody = require('./validateBody');
+
+module.exports = validateBody;
