+++ routes/api/contacts.js
@@ -3,14 +3,17 @@ const express = require("express");
 const router = express.Router();
 const ctrl = require("../../controllers/contacts");
 
+const { validateBody } = require('../../middelwares');
+const schemas = require('../../schemas/contacts')
+
 router.get("/", ctrl.getAll);
 
 router.get("/:contactId", ctrl.getById);
 
-router.post("/", ctrl.add);
+router.post("/", validateBody(schemas.addShema), ctrl.add);
 
 router.delete("/:contactId", ctrl.deleteById);
 
-router.put("/:contactId", ctrl.updateById);
+router.put("/:contactId", validateBody(schemas.addShema), ctrl.updateById);
 
 module.exports = router;
