+++ schemas/contacts.js
@@ -1,0 +1,10 @@
+const Joi = require("joi");
+
+
+const addShema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().required(),
+  phone: Joi.string().required(),
+});
+
+module.exports = {addShema};
